reorganized functions to ensure commands/config.rs only contains commands that modify config
