+++ src-tauri/src/main.rs
@@ -161,13 +161,9 @@ fn main() {
       commands::cache::get_mod_sources_data,
       commands::cache::refresh_mod_sources,
       commands::config::cleanup_enabled_texture_packs,
-      commands::config::delete_old_data_directory,
       commands::config::does_active_tooling_version_meet_minimum,
       commands::config::does_active_tooling_version_support_game,
-      commands::config::has_old_data_directory,
       commands::config::is_avx_requirement_met,
-      commands::config::is_diskspace_requirement_met,
-      commands::config::is_minimum_vcc_runtime_installed,
       commands::config::is_opengl_requirement_met,
       commands::config::reset_to_defaults,
       commands::config::set_enabled_texture_packs,
@@ -200,6 +196,10 @@ fn main() {
       commands::logging::frontend_log,
       commands::support::generate_support_package,
       commands::util::path_exists,
+      commands::util::delete_old_data_directory,
+      commands::util::is_diskspace_requirement_met,
+      commands::util::is_minimum_vcc_runtime_installed,
+      commands::util::has_old_data_directory,
       commands::versions::download_version,
       commands::versions::ensure_active_version_still_exists,
       commands::versions::go_to_version_folder,
